backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
django-filter==24.3
drf-spectacular==0.29.0